// lib/generated/LineIndexEntry.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LineMode } from "./LineMode";

export type LineIndexEntry = { broj: string, mode: LineMode, isNight: boolean, terminals: Array<string>, dailyDepartures: number, peakHeadwayMin: number | null, avgHeadwayMin: number | null, };